fix: wire globe_constellation flag into actual homepage + clean up flag categories
The globe flag was toggled in the old ConstellationHero but the active
homepage uses 4 segment variants (HomeAnonymous, HomeCitizen, HomeDRep,
HomeSPO) that all hardcoded GovernanceConstellation. Extract a shared
ConstellationScene wrapper that checks the flag and renders the correct
variant.

Also: delete dead civica_frontend flag, normalize all flag categories
to lowercase consistent grouping (citizen, governance, scoring, spo,
visualization).
